requirements.txt
tweepy
pandas
python-dotenv
textblob
openpyxl
pymongo
unicode
unidecode